--- src/store/state.ts
@@ -1,5 +1,0 @@
-import { MoviesState } from './reducer';
-
-export interface RootState {
-    movies: MoviesState;
-}
